Separation of the code in different sub code. espnow_src.c for wireless and UltrasonicSensor.c for the sensor
