+++ ESPNOW/TestEspNow_ESP32/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "espnow_example_main.c"
-                    INCLUDE_DIRS ".")
+idf_component_register(SRCS "UltrasonicSensor.c" "espnow_src.c" "espnow_example_main.c"
+                    INCLUDE_DIRS "." "include")
